use doubles on tag methods so we don't have to add F to inputs

+++ src/main/java/me/dags/noise/Module.java
@@ -188,8 +188,8 @@ public interface Module {
         return new TaggedBlend<>(this, s0, s1);
     }
 
-    default <T> Tagged<T> tagSelect(Tagged<T> s0, Tagged<T> s1, float lower, float upper, float falloff) {
-        return new TaggedSelect<>(this, s0, s1, lower, upper, falloff);
+    default <T> Tagged<T> tagSelect(Tagged<T> s0, Tagged<T> s1, double lower, double upper, double falloff) {
+        return new TaggedSelect<>(this, s0, s1, (float) lower, (float) upper, (float) falloff);
     }
 
     default void save(Path path, String... nodePath) {
